refactor(sigil): extract context menu snapshot projection
